Update
Expanded WMS for warehouse consideration and various operations. Removed IItem from Inventory Management (IUnit is currently replacing IItem). Added employee, location and schedule interfaces to Common.

+++ ERP/ERP/Inventory_Management/Items/IOrder.cs
@@ -8,7 +8,7 @@ namespace InventoryManagement.Items
 {
     public interface IOrder
     {
-        int Quantity { get; set; }
+        double Quantity { get; set; }
         double Price { get; set; }
         double Total { get; set; }
     }
